requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
confluent-kafka==1.4.1
elasticsearch==7.7.0
idna==2.9
python-dotenv==0.13.0
python-settings==0.2.1
requests==2.23.0
urllib3==1.25.9
